requirements.txt
clickclick>=0.9
easydict
keyring
keyrings.alt
PyYAML>=3.11
requests>=2.2.1
stups-zign
